move to suggested screens and tasks
