+++ x-pack/test/security_solution_cypress/cypress/screens/asset_inventory/asset_inventory_onboarding_aws.cy.ts
@@ -24,7 +24,7 @@ import {
   findPolicyLink,
   saveIntegration,
   selectPolicyForEditing,
-} from './constants';
+} from '../../tasks/asset_inventory/common';
 
 const ROLE_ARN = getDataTestSubjectSelector('awsRoleArnInput');
 const ACCESS_KEY = getDataTestSubjectSelector('awsDirectAccessKeyId');
