+++ x-pack/test/security_solution_cypress/cypress/screens/asset_inventory/asset_inventory_onboarding_azure.cy.ts
@@ -25,7 +25,7 @@ import {
   saveIntegration,
   selectPolicyForEditing,
   selectCloudProvider,
-} from './constants';
+} from '../../tasks/asset_inventory/common';
 
 const AZURE_ORGANIZATION_ACCOUNT_TEST_ID = '#organization-account';
 const AZURE_SINGLE_ACCOUNT_TEST_ID = '#single-account';
