+++ x-pack/test/security_solution_cypress/cypress/screens/asset_inventory/asset_inventory_onboarding_gcp.cy.ts
@@ -27,7 +27,7 @@ import {
   saveIntegration,
   selectCloudProvider,
   selectPolicyForEditing,
-} from './constants';
+} from '../../tasks/asset_inventory/common';
 
 const GCP_CLOUDSHELL_TEST_ID = `${getDataTestSubjectSelector(
   'gcpGoogleCloudShellOptionTestId'
