+++ x-pack/test/security_solution_cypress/cypress/screens/asset_inventory/asset_inventory_page.cy.ts
