+++ x-pack/test/security_solution_cypress/cypress/tasks/asset_inventory/common.ts
